// lib/interfaces/key.interface.ts
export enum CylinderKey {
  AFRI_GAS = 'AFRI_GAS',
  ALFA_GAS = 'ALFA_GAS',
  AMAAN_GAS = 'AMAAN_GAS',
  CAPITAL_GAS = 'CAPITAL_GAS',
  CITY_GAS = 'CITY_GAS',
  DEL_GAS = 'DEL_GAS',
  FAST_GAS = 'FAST_GAS',
  FUTURE_GAS = 'FUTURE_GAS',
  G_GAS = 'G_GAS',
  GASKY_GAS = 'GASKY_GAS',
  GULF_PETROCHEM_GAS = 'GULF_PETROCHEM_GAS',
  HANDI_GAS = 'HANDI_GAS',
  HASHI_GAS = 'HASHI_GAS',
  HASS_GAS = 'HASS_GAS',
  HOME_GAS = 'HOME_GAS',
  HUNKAR_GAS = 'HUNKAR_GAS',
  JAMII_GAS = 'JAMII_GAS',
  K_GAS = 'K_GAS',
  KAPRI_GAS = 'KAPRI_GAS',
  KERRY_GAS = 'KERRY_GAS',
  LAKE_GAS = 'LAKE_GAS',
  MEN_GAS = 'MEN_GAS',
  MID_GAS = 'MID_GAS',
  MOTO_GAS = 'MOTO_GAS',
  MPISHI_GAS = 'MPISHI_GAS',
  OILCOM_GAS = 'OILCOM_GAS',
  ORANGE_GAS = 'ORANGE_GAS',
  ORYX_GAS = 'ORYX_GAS',
  P_GAS = 'P_GAS',
  PET_GAS = 'PET_GAS',
  PRO_GAS = 'PRO_GAS',
  SAFARI_GAS = 'SAFARI_GAS',
  SAFE_GAS = 'SAFE_GAS',
  SALAMA_GAS = 'SALAMA_GAS',
  SEA_GAS = 'SEA_GAS',
  SUPA_GAS = 'SUPA_GAS',
  T_GAS = 'T_GAS',
  TEX_GAS = 'TEX_GAS',
  TOSHA_GAS = 'TOSHA_GAS',
  TOTAL_GAS = 'TOTAL_GAS',
  VENN_GAS = 'VENN_GAS',
}